move functions to read_transit
